requirements.txt
# Core dependencies (auto-installed with pip install -e .)
numpy>=1.20.0
pandas>=1.3.0
rdkit>=2022.3.1
tqdm>=4.60.0
ase>=3.22.0
pyyaml>=6.0
fairchem-core

# Install separately via conda
# openbabel>=3.1.0